Fill color files refactor
