+++ lib/assets/javascripts/builder/components/form-components/editors/fill/fill-color/fill-color-by-value-view.js
@@ -2,7 +2,7 @@ var _ = require('underscore');
 var CoreView = require('backbone/core-view');
 
 var InputCollection = require('builder/components/form-components/editors/fill/fill-color/fill-color-input-collection');
-var InputColorByValueView = require('builder/components/form-components/editors/fill/fill-color/input-color-by-value/input-color-by-value');
+var InputColorByValueView = require('builder/components/form-components/editors/fill/fill-color/inputs/input-color-by-value');
 
 var FillTemplate = require('builder/components/form-components/editors/fill/fill-template.tpl');
 var FillDialogModel = require('builder/components/form-components/editors/fill/fill-dialog-model');
