+++ lib/assets/javascripts/builder/components/form-components/editors/fill/fill-color/fill-color-solid-view.js
@@ -4,8 +4,8 @@ var FillDialogModel = require('builder/components/form-components/editors/fill/f
 var FillDialogView = require('builder/components/form-components/editors/fill/fill-dialog');
 
 var InputCollection = require('builder/components/form-components/editors/fill/fill-color/fill-color-input-collection');
-var InputColorSolidView = require('builder/components/form-components/editors/fill/fill-color/input-color-solid/input-color-solid');
-var InputImageView = require('builder/components/form-components/editors/fill/fill-color/input-img/input-img');
+var InputColorSolidView = require('builder/components/form-components/editors/fill/fill-color/inputs/input-color-solid');
+var InputImageView = require('builder/components/form-components/editors/fill/fill-color/inputs/input-image');
 
 var FillTemplate = require('builder/components/form-components/editors/fill/fill-template.tpl');
 var PopupManager = require('builder/components/popup-manager');
