+++ lib/assets/javascripts/builder/components/form-components/editors/fill/fill-color/inputs/input-color-by-value.js
@@ -3,11 +3,22 @@ var $ = require('jquery');
 
 var CoreView = require('backbone/core-view');
 
-var ColorPickerView = require('builder/components/form-components/editors/fill/color-picker/color-picker');
+var InputColorValueContentView = require('builder/components/form-components/editors/fill/input-color/input-color-value-content-view');
 var TipsyTooltipView = require('builder/components/tipsy-tooltip-view');
-var InputColorTemplate = require('builder/components/form-components/editors/fill/fill-color/fill-color-input.tpl');
+var InputColorTemplate = require('builder/components/form-components/editors/fill/fill-color/inputs/input-color.tpl');
 
 var Utils = require('builder/helpers/utils');
+
+var checkAndBuildOpts = require('builder/helpers/required-opts');
+
+var REQUIRED_OPTS = [
+  'columns',
+  'query',
+  'configModel',
+  'userModel',
+  'editorAttrs',
+  'modals'
+];
 
 module.exports = CoreView.extend({
   tagName: 'li',
@@ -32,11 +43,7 @@ module.exports = CoreView.extend({
 
     if (!options.columns) throw new Error('columns is required');
 
-    this._columns = options.columns;
-    this._configModel = options.configModel;
-    this._userModel = options.userModel;
-    this._modals = options.modals;
-    this._query = options.query;
+    checkAndBuildOpts(options, REQUIRED_OPTS, this);
 
     this._initBinds();
   },
@@ -92,15 +99,17 @@ module.exports = CoreView.extend({
   },
 
   _createContentView: function () {
-    this._colorPickerView = new ColorPickerView({
-      value: this.model.get('fixed'),
-      opacity: this.model.get('opacity'),
-      disableOpacity: this._disableOpacity || false
+    return new InputColorValueContentView({
+      model: this.model,
+      columns: this._columns,
+      configModel: this._configModel,
+      categorizeColumns: this._categorizeColumns,
+      imageEnabled: this._imageEnabled,
+      userModel: this._userModel,
+      modals: this._modals,
+      hideTabs: this._hideTabs,
+      query: this._query
     });
-
-    this._initColorPickerViewBindings();
-
-    return this._colorPickerView;
   },
 
   _initColorPickerViewBindings: function () {
