+++ lib/assets/javascripts/builder/components/form-components/editors/fill/fill-color/inputs/input-color.tpl
