+++ lib/assets/javascripts/builder/components/form-components/editors/fill/fill-color/inputs/input-image.js
@@ -3,7 +3,7 @@ var $ = require('jquery');
 
 var CoreView = require('backbone/core-view');
 
-var InputImageTemplate = require('./input-img.tpl');
+var InputImageTemplate = require('./input-image.tpl');
 var ImageLoaderView = require('builder/components/img-loader-view');
 var InputColorFileView = require('builder/components/form-components/editors/fill/input-color/input-color-file-view');
 var TipsyTooltipView = require('builder/components/tipsy-tooltip-view');
